no generic wrapping of java objects; avoid shorts in the js interfaces

+++ src/main/java/org/htmlunit/activex/javascript/msxml/XMLDOMNode.java
@@ -227,7 +227,7 @@ public class XMLDOMNode extends MSXMLScriptable {
      * @return the XML Document Object Model (DOM) node type
      */
     @JsxGetter
-    public short getNodeType() {
+    public int getNodeType() {
         final DomNode domNode = getDomNodeOrDie();
         return domNode.getNodeType();
     }
