+++ src/main/java/org/htmlunit/javascript/host/dom/DOMException.java
@@ -40,85 +40,85 @@ import org.htmlunit.javascript.configuration.JsxGetter;
 public class DOMException extends HtmlUnitScriptable {
     /** If the specified range of text does not fit into a DOMString. */
     @JsxConstant
-    public static final short DOMSTRING_SIZE_ERR = org.w3c.dom.DOMException.DOMSTRING_SIZE_ERR;
+    public static final int DOMSTRING_SIZE_ERR = org.w3c.dom.DOMException.DOMSTRING_SIZE_ERR;
     /** If any node is inserted somewhere it doesn't belong. */
     @JsxConstant
-    public static final short HIERARCHY_REQUEST_ERR = org.w3c.dom.DOMException.HIERARCHY_REQUEST_ERR;
+    public static final int HIERARCHY_REQUEST_ERR = org.w3c.dom.DOMException.HIERARCHY_REQUEST_ERR;
     /** If index or size is negative, or greater than the allowed value. */
     @JsxConstant
-    public static final short INDEX_SIZE_ERR = org.w3c.dom.DOMException.INDEX_SIZE_ERR;
+    public static final int INDEX_SIZE_ERR = org.w3c.dom.DOMException.INDEX_SIZE_ERR;
     /** If an attempt is made to add an attribute that is already in use elsewhere. */
     @JsxConstant
-    public static final short INUSE_ATTRIBUTE_ERR = org.w3c.dom.DOMException.INUSE_ATTRIBUTE_ERR;
+    public static final int INUSE_ATTRIBUTE_ERR = org.w3c.dom.DOMException.INUSE_ATTRIBUTE_ERR;
     /** If a parameter or an operation is not supported by the underlying object. */
     @JsxConstant
-    public static final short INVALID_ACCESS_ERR = org.w3c.dom.DOMException.INVALID_ACCESS_ERR;
+    public static final int INVALID_ACCESS_ERR = org.w3c.dom.DOMException.INVALID_ACCESS_ERR;
     /** If an invalid or illegal character is specified, such as in a name. */
     @JsxConstant
-    public static final short INVALID_CHARACTER_ERR = org.w3c.dom.DOMException.INVALID_CHARACTER_ERR;
+    public static final int INVALID_CHARACTER_ERR = org.w3c.dom.DOMException.INVALID_CHARACTER_ERR;
     /** If an attempt is made to modify the type of the underlying object. */
     @JsxConstant
-    public static final short INVALID_MODIFICATION_ERR = org.w3c.dom.DOMException.INVALID_MODIFICATION_ERR;
+    public static final int INVALID_MODIFICATION_ERR = org.w3c.dom.DOMException.INVALID_MODIFICATION_ERR;
     /** If an attempt is made to use an object that is not, or is no longer, usable. */
     @JsxConstant
-    public static final short INVALID_STATE_ERR = org.w3c.dom.DOMException.INVALID_STATE_ERR;
+    public static final int INVALID_STATE_ERR = org.w3c.dom.DOMException.INVALID_STATE_ERR;
     /** If an attempt is made to create or change an object in a way which is incorrect with regard to namespaces. */
     @JsxConstant
-    public static final short NAMESPACE_ERR = org.w3c.dom.DOMException.NAMESPACE_ERR;
+    public static final int NAMESPACE_ERR = org.w3c.dom.DOMException.NAMESPACE_ERR;
     /** If data is specified for a node which does not support data. */
     @JsxConstant
-    public static final short NO_DATA_ALLOWED_ERR = org.w3c.dom.DOMException.NO_DATA_ALLOWED_ERR;
+    public static final int NO_DATA_ALLOWED_ERR = org.w3c.dom.DOMException.NO_DATA_ALLOWED_ERR;
     /** If an attempt is made to modify an object where modifications are not allowed. */
     @JsxConstant
-    public static final short NO_MODIFICATION_ALLOWED_ERR = org.w3c.dom.DOMException.NO_MODIFICATION_ALLOWED_ERR;
+    public static final int NO_MODIFICATION_ALLOWED_ERR = org.w3c.dom.DOMException.NO_MODIFICATION_ALLOWED_ERR;
     /** If an attempt is made to reference a node in a context where it does not exist. */
     @JsxConstant
-    public static final short NOT_FOUND_ERR = org.w3c.dom.DOMException.NOT_FOUND_ERR;
+    public static final int NOT_FOUND_ERR = org.w3c.dom.DOMException.NOT_FOUND_ERR;
     /** If the implementation does not support the requested type of object or operation. */
     @JsxConstant
-    public static final short NOT_SUPPORTED_ERR = org.w3c.dom.DOMException.NOT_SUPPORTED_ERR;
+    public static final int NOT_SUPPORTED_ERR = org.w3c.dom.DOMException.NOT_SUPPORTED_ERR;
     /** If an invalid or illegal string is specified. */
     @JsxConstant
-    public static final short SYNTAX_ERR = org.w3c.dom.DOMException.SYNTAX_ERR;
+    public static final int SYNTAX_ERR = org.w3c.dom.DOMException.SYNTAX_ERR;
     /** If a node is used in a different document than the one that created it (that doesn't support it). */
     @JsxConstant
-    public static final short WRONG_DOCUMENT_ERR = org.w3c.dom.DOMException.WRONG_DOCUMENT_ERR;
+    public static final int WRONG_DOCUMENT_ERR = org.w3c.dom.DOMException.WRONG_DOCUMENT_ERR;
     /** If a call to a method would make the {@code Node} invalid with respect to "partial validity". */
     @JsxConstant
-    public static final short VALIDATION_ERR = org.w3c.dom.DOMException.VALIDATION_ERR;
+    public static final int VALIDATION_ERR = org.w3c.dom.DOMException.VALIDATION_ERR;
     /** If the type of an object is incompatible with the expected type of the parameter. */
     @JsxConstant
-    public static final short TYPE_MISMATCH_ERR = org.w3c.dom.DOMException.TYPE_MISMATCH_ERR;
+    public static final int TYPE_MISMATCH_ERR = org.w3c.dom.DOMException.TYPE_MISMATCH_ERR;
     /** Security error. */
     @JsxConstant
-    public static final short SECURITY_ERR = 18;
+    public static final int SECURITY_ERR = 18;
     /** Network error. */
     @JsxConstant
-    public static final short NETWORK_ERR = 19;
+    public static final int NETWORK_ERR = 19;
     /** Abort error. */
     @JsxConstant
-    public static final short ABORT_ERR = 20;
+    public static final int ABORT_ERR = 20;
     /** URL mismatch error. */
     @JsxConstant
-    public static final short URL_MISMATCH_ERR = 21;
+    public static final int URL_MISMATCH_ERR = 21;
     /** Quota exceeded error. */
     @JsxConstant
-    public static final short QUOTA_EXCEEDED_ERR = 22;
+    public static final int QUOTA_EXCEEDED_ERR = 22;
     /** Timeout error. */
     @JsxConstant
-    public static final short TIMEOUT_ERR = 23;
+    public static final int TIMEOUT_ERR = 23;
     /** Invalid node type error. */
     @JsxConstant
-    public static final short INVALID_NODE_TYPE_ERR = 24;
+    public static final int INVALID_NODE_TYPE_ERR = 24;
     /** Data clone error. */
     @JsxConstant
-    public static final short DATA_CLONE_ERR = 25;
+    public static final int DATA_CLONE_ERR = 25;
     /** Parse error. */
     @JsxConstant(IE)
-    public static final short PARSE_ERR = 81;
+    public static final int PARSE_ERR = 81;
     /** Serialize error. */
     @JsxConstant(IE)
-    public static final short SERIALIZE_ERR = 82;
+    public static final int SERIALIZE_ERR = 82;
 
     private final short code_;
     private final String message_;
