+++ src/main/java/org/htmlunit/javascript/host/event/EventSource.java
@@ -34,15 +34,15 @@ public class EventSource extends EventTarget {
 
     /** The connection is being established. */
     @JsxConstant
-    public static final short CONNECTING = 0;
+    public static final int CONNECTING = 0;
 
     /** The connection is open and dispatching events. */
     @JsxConstant
-    public static final short OPEN = 1;
+    public static final int OPEN = 1;
 
     /** The connection is not being established, has been closed or there was a fatal error. */
     @JsxConstant
-    public static final short CLOSED = 2;
+    public static final int CLOSED = 2;
 
     /**
      * Default constructor.
