+++ src/main/java/org/htmlunit/javascript/host/geo/PositionError.java
@@ -24,19 +24,20 @@ import org.htmlunit.javascript.configuration.JsxConstant;
  * A JavaScript object for {@code PositionError}.
  *
  * @author Ahmed Ashour
+ * @author Ronald Brill
  */
 @JsxClass(IE)
 public class PositionError extends HtmlUnitScriptable {
 
     /** The constant for {@code PERMISSION_DENIED}. */
     @JsxConstant
-    public static final short PERMISSION_DENIED = 1;
+    public static final int PERMISSION_DENIED = 1;
     /** The constant for {@code POSITION_UNAVAILABLE}. */
     @JsxConstant
-    public static final short POSITION_UNAVAILABLE = 2;
+    public static final int POSITION_UNAVAILABLE = 2;
     /** The constant for {@code TIMEOUT}. */
     @JsxConstant
-    public static final short TIMEOUT = 3;
+    public static final int TIMEOUT = 3;
 
     /**
      * Default constructor.
