+++ src/main/java/org/htmlunit/javascript/host/html/HTMLMediaElement.java
@@ -44,50 +44,50 @@ public class HTMLMediaElement extends HTMLElement {
      * No information is available about the media resource.
      */
     @JsxConstant
-    public static final short HAVE_NOTHING = 0;
+    public static final int HAVE_NOTHING = 0;
 
     /**
      * Enough of the media resource has been retrieved that the metadata attributes are initialized.
      * Seeking will no longer raise an exception.
      */
     @JsxConstant
-    public static final short HAVE_METADATA = 1;
+    public static final int HAVE_METADATA = 1;
 
     /**
      * Data is available for the current playback position, but not enough to actually play more than one frame.
      */
     @JsxConstant
-    public static final short HAVE_CURRENT_DATA = 2;
+    public static final int HAVE_CURRENT_DATA = 2;
 
     /**
      * Data for the current playback position as well as for at least a little bit of time
      * into the future is available (in other words, at least two frames of video, for example).
      */
     @JsxConstant
-    public static final short HAVE_FUTURE_DATA = 3;
+    public static final int HAVE_FUTURE_DATA = 3;
 
     /**
      * Enough data is available—and the download rate is high enough—that the media
      * can be played through to the end without interruption.
      */
     @JsxConstant
-    public static final short HAVE_ENOUGH_DATA = 4;
+    public static final int HAVE_ENOUGH_DATA = 4;
 
     /** There is no data yet. */
     @JsxConstant
-    public static final short NETWORK_EMPTY = 0;
+    public static final int NETWORK_EMPTY = 0;
 
     /** Network is idle. */
     @JsxConstant
-    public static final short NETWORK_IDLE = 1;
+    public static final int NETWORK_IDLE = 1;
 
     /** The media is loading. */
     @JsxConstant
-    public static final short NETWORK_LOADING = 2;
+    public static final int NETWORK_LOADING = 2;
 
     /** There is no source. */
     @JsxConstant
-    public static final short NETWORK_NO_SOURCE = 3;
+    public static final int NETWORK_NO_SOURCE = 3;
 
     /**
      * Creates an instance.
@@ -147,7 +147,7 @@ public class HTMLMediaElement extends HTMLElement {
      */
     @JsxGetter
     @Override
-    public short getNodeType() {
+    public int getNodeType() {
         final HtmlMedia element = (HtmlMedia) getDomNodeOrNull();
         if (element == null) {
             return Node.ELEMENT_NODE;
